requirements.txt
streamlit
pandas
google-generativeai
fpdf2
